Move grunt config to config dir
