+++ config/grunt/custom-tasks/convert-pot-to-wp.js
