+++ config/grunt/custom-tasks/ensure-rc-dependencies.js
