+++ config/grunt/custom-tasks/remove-vendor-prefixed-folder.js
