+++ config/grunt/tasks/convert-pot-to-wp.js
