+++ config/grunt/tasks/makepot.js
