+++ config/grunt/tasks/notify-slack.js
