+++ config/grunt/tasks/po2json.js
